// src/charting_library/charting_library.min.d.ts
export declare type LanguageCode =
  | 'ar'
  | 'zh'
  | 'cs'
  | 'da_DK'
  | 'nl_NL'
  | 'en'
  | 'et_EE'
  | 'fr'
  | 'de'
  | 'el'
  | 'he_IL'
  | 'hu_HU'
  | 'id_ID'
  | 'it'
  | 'ja'
  | 'ko'
  | 'fa'
  | 'pl'
  | 'pt'
  | 'ro'
  | 'ru'
  | 'sk_SK'
  | 'es'
  | 'sv'
  | 'th'
  | 'tr'
  | 'vi';
// ...
export declare type StudyOverrideValueType = string | number | boolean;
export interface StudyOverrides {
  [key: string]: StudyOverrideValueType;
}
// ...
export declare type ResolutionString = string;
// ...
export declare type OnReadyCallback = (
  configuration: DatafeedConfiguration
) => void;
export interface IExternalDatafeed {
  onReady(callback: OnReadyCallback): void;
}
// ...
export declare type QuotesCallback = (data: QuoteData[]) => void;
export interface IDatafeedQuotesApi {
  getQuotes(
    symbols: string[],
    onDataCallback: QuotesCallback,
    onErrorCallback: (msg: string) => void
  ): void;
  subscribeQuotes(
    symbols: string[],
    fastSymbols: string[],
    onRealtimeCallback: QuotesCallback,
    listenerGUID: string
  ): void;
  unsubscribeQuotes(listenerGUID: string): void;
}
export declare type CustomTimezones =
  | 'America/New_York'
  | 'America/Los_Angeles'
  | 'America/Chicago'
  | 'America/Phoenix'
  | 'America/Toronto'
  | 'America/Vancouver'
  | 'America/Argentina/Buenos_Aires'
  | 'America/El_Salvador'
  | 'America/Sao_Paulo'
  | 'America/Bogota'
  | 'America/Caracas'
  | 'Europe/Moscow'
  | 'Europe/Athens'
  | 'Europe/Berlin'
  | 'Europe/London'
  | 'Europe/Madrid'
  | 'Europe/Paris'
  | 'Europe/Rome'
  | 'Europe/Warsaw'
  | 'Europe/Istanbul'
  | 'Europe/Zurich'
  | 'Australia/Sydney'
  | 'Australia/Brisbane'
  | 'Australia/Adelaide'
  | 'Australia/ACT'
  | 'Asia/Almaty'
  | 'Asia/Ashkhabad'
  | 'Asia/Tokyo'
  | 'Asia/Taipei'
  | 'Asia/Singapore'
  | 'Asia/Shanghai'
  | 'Asia/Seoul'
  | 'Asia/Tehran'
  | 'Asia/Dubai'
  | 'Asia/Kolkata'
  | 'Asia/Hong_Kong'
  | 'Asia/Bangkok'
  | 'Pacific/Auckland'
  | 'Pacific/Chatham'
  | 'Pacific/Fakaofo'
  | 'Pacific/Honolulu'
  | 'America/Mexico_City'
  | 'Africa/Johannesburg'
  | 'Asia/Kathmandu'
  | 'US/Mountain';
export declare type Timezone = 'UTC' | CustomTimezones;
export interface LibrarySymbolInfo {
  /**
   * Symbol Name
   */
  name: string;
  full_name: string;
  base_name?: [string];
  /**
   * Unique symbol id
   */
  ticker?: string;
  description: string;
  type: string;
  /**
   * @example "1700-0200"
   */
  session: string;
  /**
   * Traded exchange
   * @example "NYSE"
   */
  exchange: string;
  listed_exchange: string;
  timezone: Timezone;
  /**
   * Code (Tick)
   * @example 8/16/.../256 (1/8/100 1/16/100 ... 1/256/100) or 1/10/.../10000000 (1 0.1 ... 0.0000001)
   */
  pricescale: number;
  /**
   * The number of units that make up one tick.
   * @example For example, U.S. equities are quotes in decimals, and tick in decimals, and can go up +/- .01. So the tick increment is 1. But the e-mini S&P futures contract, though quoted in decimals, goes up in .25 increments, so the tick increment is 25. (see also Tick Size)
   */
  minmov: number;
  fractional?: boolean;
  /**
   * @example Quarters of 1/32: pricescale=128, minmovement=1, minmovement2=4
   */
  minmove2?: number;
  /**
   * false if DWM only
   */
  has_intraday?: boolean;
  /**
   * An array of resolutions which should be enabled in resolutions picker for this symbol.
   */
  supported_resolutions: ResolutionString[];
  /**
   * @example (for ex.: "1,5,60") - only these resolutions will be requested, all others will be built using them if possible
   */
  intraday_multipliers?: string[];
  has_seconds?: boolean;
  /**
   * It is an array containing seconds resolutions (in seconds without a postfix) the datafeed builds by itself.
   */
  seconds_multipliers?: string[];
  has_daily?: boolean;
  has_weekly_and_monthly?: boolean;
  has_empty_bars?: boolean;
  force_session_rebuild?: boolean;
  has_no_volume?: boolean;
  /**
   * Integer showing typical volume value decimal places for this symbol
   */
  volume_precision?: number;
  data_status?: 'streaming' | 'endofday' | 'pulsed' | 'delayed_streaming';
  /**
   * Boolean showing whether this symbol is expired futures contract or not.
   */
  expired?: boolean;
  /**
   * Unix timestamp of expiration date.
   */
  expiration_date?: number;
  sector?: string;
  industry?: string;
  currency_code?: string;
}
// ...
export interface Mark {
  id: string | number;
  time: number;
  color: MarkConstColors | MarkCustomColor;
  text: string;
  label: string;
  labelFontColor: string;
  minSize: number;
}
export interface TimescaleMark {
  id: string | number;
  time: number;
  color: MarkConstColors | string;
  label: string;
  tooltip: string[];
}
export declare type ResolutionBackValues = 'D' | 'M';
export interface HistoryDepth {
  resolutionBack: ResolutionBackValues;
  intervalBack: number;
}
export declare type SearchSymbolsCallback = (
  items: SearchSymbolResultItem[]
) => void;
export declare type ResolveCallback = (symbolInfo: LibrarySymbolInfo) => void;
export declare type HistoryCallback = (
  bars: Bar[],
  meta: HistoryMetadata
) => void;
export declare type SubscribeBarsCallback = (bar: Bar) => void;
export declare type GetMarksCallback<T> = (marks: T[]) => void;
export declare type ServerTimeCallback = (serverTime: number) => void;
export declare type DomeCallback = (data: DOMData) => void;
export declare type ErrorCallback = (reason: string) => void;
export interface IDatafeedChartApi {
  calculateHistoryDepth?(
    resolution: ResolutionString,
    resolutionBack: ResolutionBackValues,
    intervalBack: number
  ): HistoryDepth | undefined;
  getMarks?(
    symbolInfo: LibrarySymbolInfo,
    startDate: number,
    endDate: number,
    onDataCallback: GetMarksCallback<Mark>,
    resolution: ResolutionString
  ): void;
  getTimescaleMarks?(
    symbolInfo: LibrarySymbolInfo,
    startDate: number,
    endDate: number,
    onDataCallback: GetMarksCallback<TimescaleMark>,
    resolution: ResolutionString
  ): void;
  /**
   * This function is called if configuration flag supports_time is set to true when chart needs to know the server time.
   * The charting library expects callback to be called once.
   * The time is provided without milliseconds. Example: 1445324591. It is used to display Countdown on the price scale.
   */
  getServerTime?(callback: ServerTimeCallback): void;
  searchSymbols(
    userInput: string,
    exchange: string,
    symbolType: string,
    onResult: SearchSymbolsCallback
  ): void;
  resolveSymbol(
    symbolName: string,
    onResolve: ResolveCallback,
    onError: ErrorCallback
  ): void;
  getBars(
    symbolInfo: LibrarySymbolInfo,
    resolution: ResolutionString,
    rangeStartDate: number,
    rangeEndDate: number,
    onResult: HistoryCallback,
    onError: ErrorCallback,
    isFirstCall: boolean
  ): void;
  subscribeBars(
    symbolInfo: LibrarySymbolInfo,
    resolution: ResolutionString,
    onTick: SubscribeBarsCallback,
    listenerGuid: string,
    onResetCacheNeededCallback: () => void
  ): void;
  unsubscribeBars(listenerGuid: string): void;
  subscribeDepth?(
    symbolInfo: LibrarySymbolInfo,
    callback: DomeCallback
  ): string;
  unsubscribeDepth?(subscriberUID: string): void;
}
export interface ChartMetaInfo {
  id: string;
  name: string;
  symbol: string;
  resolution: ResolutionString;
  timestamp: number;
}
export interface ChartData {
  id: string;
  name: string;
  symbol: string;
  resolution: ResolutionString;
  content: string;
}
export interface StudyTemplateMetaInfo {
  name: string;
}
export interface StudyTemplateData {
  name: string;
  content: string;
}
export interface IExternalSaveLoadAdapter {
  getAllCharts(): Promise<ChartMetaInfo[]>;
  removeChart(chartId: string): Promise<void>;
  saveChart(chartData: ChartData): Promise<string>;
  getChartContent(chartId: string): Promise<string>;
  getAllStudyTemplates(): Promise<StudyTemplateMetaInfo[]>;
  removeStudyTemplate(studyTemplateInfo: StudyTemplateMetaInfo): Promise<void>;
  saveStudyTemplate(studyTemplateData: StudyTemplateData): Promise<void>;
  getStudyTemplateContent(
    studyTemplateInfo: StudyTemplateMetaInfo
  ): Promise<string>;
}
export interface RestBrokerMetaInfo {
  url: string;
  access_token: string;
}
export interface AccessListItem {
  name: string;
  grayed?: boolean;
}
export interface AccessList {
  type: 'black' | 'white';
  tools: AccessListItem[];
}
export interface NumericFormattingParams {
  decimal_sign: string;
}
export declare type AvailableSaveloadVersions = '1.0' | '1.1';
export interface Overrides {
  [key: string]: string | number | boolean;
}
// ...
export interface CustomFormatter {
  format(date: Date): string;
  formatLocal(date: Date): string;
}
export interface CustomFormatters {
  timeFormatter: CustomFormatter;
  dateFormatter: CustomFormatter;
}
export interface TimeFrameItem {
  text: string;
  resolution: ResolutionString;
  description?: string;
  title?: string;
}
export interface Favorites {
  intervals: ResolutionString[];
  chartTypes: string[];
}
export interface NewsItem {
  fullDescription: string;
  link?: string;
  published: number;
  shortDescription?: string;
  source: string;
  title: string;
}
export interface LoadingScreenOptions {
  foregroundColor?: string;
  backgroundColor?: string;
}
export interface InitialSettingsMap {
  [key: string]: string;
}
export interface ISettingsAdapter {
  initialSettings?: InitialSettingsMap;
  setValue(key: string, value: string): void;
  removeValue(key: string): void;
}
export declare type IBasicDataFeed = IDatafeedChartApi & IExternalDatafeed;
export interface ChartingLibraryWidgetOptions {
  container_id: string;
  datafeed: IBasicDataFeed | (IBasicDataFeed & IDatafeedQuotesApi);
  interval: ResolutionString;
  symbol: string;
  auto_save_delay?: number;
  autosize?: boolean;
  debug?: boolean;
  disabled_features?: string[];
  drawings_access?: AccessList;
  enabled_features?: string[];
  fullscreen?: boolean;
  height?: number;
  library_path?: string;
  locale: LanguageCode;
  numeric_formatting?: NumericFormattingParams;
  saved_data?: object;
  studies_access?: AccessList;
  study_count_limit?: number;
  symbol_search_request_delay?: number;
  timeframe?: string;
  timezone?: 'exchange' | Timezone;
  toolbar_bg?: string;
  width?: number;
  charts_storage_url?: string;
  charts_storage_api_version?: AvailableSaveloadVersions;
  client_id?: string;
  user_id?: string;
  load_last_chart?: boolean;
  studies_overrides?: StudyOverrides;
  customFormatters?: CustomFormatters;
  overrides?: Overrides;
  snapshot_url?: string;
  indicators_file_name?: string;
  preset?: 'mobile';
  time_frames?: TimeFrameItem[];
  custom_css_url?: string;
  favorites?: Favorites;
  save_load_adapter?: IExternalSaveLoadAdapter;
  loading_screen?: LoadingScreenOptions;
  settings_adapter?: ISettingsAdapter;
}
export interface TradingTerminalWidgetOptions
  extends ChartingLibraryWidgetOptions {
  brokerConfig?: SingleBrokerMetaInfo;
  restConfig?: RestBrokerMetaInfo;
  widgetbar?: WidgetBarParams;
  rss_news_feed?: RssNewsFeedParams;
  news_provider?: NewsProvider;
  brokerFactory?(
    host: IBrokerConnectionAdapterHost
  ): IBrokerWithoutRealtime | IBrokerTerminal;
}
// ...
export interface ChartingLibraryWidgetConstructor {
  new (
    options: ChartingLibraryWidgetOptions | TradingTerminalWidgetOptions
  ): IChartingLibraryWidget;
}
export declare function version(): string;
export declare function onready(callback: () => void): void;
export declare const widget: ChartingLibraryWidgetConstructor;